refactor: move cache integration tests to separate cache_test.rs
Run cache tests: HF_TOKEN=hf_xxx cargo test -p huggingface-hub --test cache_test
